requirements.txt
aiohttp
aiosignal
arabic-reshaper
asgiref
asn1crypto
attrs
certifi
cffi
chardet
charset-normalizer
click
colorama
crispy-bootstrap5
cryptography
cssselect2
discord
discord.py
dj-database-url
Django
django-crispy-forms
django-email
django-environ
django-filter
django-humanize
djangorestframework
EasyProcess
entrypoint2
frozenlist
html5lib
humanize
idna
lxml
Markdown
MouseInfo
mss
multidict
mysql
mysqlclient
numpy
opencv-contrib-python
oscrypto
Pillow
psycopg2
psycopg2-binary
PyAutoGUI
pycparser
PyGetWindow
pyHanko
pyhanko-certvalidator
PyMsgBox
pypdf
pyperclip
pypng
PyRect
pyscreenshot
PyScreeze
python-bidi
pytweening
pytz
PyYAML
qrcode
reportlab
requests
six
sqlparse
svglib
tinycss2
typing_extensions
tzdata
tzlocal
uritools
urllib3
webencodings
yarl